Authenticated routing (#375)
* Added new user context control

* Fixed b1 routes

* Completed /portal section

* Converted /tools

* Converted /admin

* Fixed admin errors

* Removed /pages

* Fxed compilation errors

* Added error handling
