+++ src/app/b1/classroom/[id]/page.tsx
