+++ src/app/b1/person/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import React, { Suspense } from "react";
+
+import { PersonInner } from "./components/PersonInner";
+
+export default function PersonPage() {
+  return <Suspense><PersonInner /></Suspense>
+}
