+++ src/app/b1/venue/[id]/components/VenueClient.tsx
