+++ src/app/b1/venue/[id]/page.tsx
