--- src/app/b1New/person/page.tsx
@@ -1,10 +1,0 @@
-"use client";
-
-import React, { Suspense } from "react";
-
-import { UserProvider } from "@/UserContext";
-import { PersonInner } from "./components/PersonInner";
-
-export default function PersonPage() {
-  return <UserProvider><Suspense><PersonInner /></Suspense></UserProvider>
-}
