+++ src/app/tools/oll/page.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { Wrapper } from "@/components/Wrapper";
+import { Suspense } from "react";
+import OllInner from "./components/OllInner";
+
+export default function CP() {
+
+  return (
+    <Wrapper>
+      <Suspense>
+        <OllInner />
+      </Suspense>
+    </Wrapper>
+  );
+}
+
