--- src/pages/_app.tsx
@@ -1,67 +1,0 @@
-import "react-activity/dist/Dots.css";
-import "@/styles/globals.css";
-import "@churchapps/apphelper/dist/components/markdownEditor/editor.css";
-import { useEffect } from "react";
-import type { AppProps } from "next/app";
-import { EnvironmentHelper } from "@/utils";
-import { AnalyticsHelper } from "@churchapps/apphelper";
-import { UserProvider } from "@/UserContext";
-
-EnvironmentHelper.init();
-
-function LessonsApp({ Component, pageProps }: AppProps) {
-
-  const location = (typeof(window) === "undefined") ? null : window.location;
-  AnalyticsHelper.init();
-  useEffect(() => { AnalyticsHelper.logPageView() }, [location]);
-
-  /*
-  useEffect(() => {
-    if (EnvironmentHelper.GoogleAnalyticsTag) {
-      const handleRouteChange = (url: string) => {
-        GoogleAnalyticsHelper.pageview(url);
-      };
-      router.events.on("routeChangeComplete", handleRouteChange);
-      return () => {
-        router.events.off("routeChangeComplete", handleRouteChange);
-      };
-    }
-  }, [router.events]);
-  */
-
-  // auto login when jwt found in cookies
-  /*
-  useEffect(() => {
-    if (cookies.jwt && !router.pathname.includes('login')) {
-      router.push("/login")
-    }
-  }, [])
-*/
-  /*
-<Head>
-{EnvironmentHelper.GoogleAnalyticsTag && (
-          <>
-            <script async src={`https://www.googletagmanager.com/gtag/js?id=${EnvironmentHelper.GoogleAnalyticsTag}`} />
-            <script
-              dangerouslySetInnerHTML={{
-                __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${EnvironmentHelper.GoogleAnalyticsTag}', {
-              page_path: window.location.pathname,
-            });
-          `,
-              }}
-            />
-          </>
-        )}
-        </Head>
-        */
-  return (
-    <UserProvider>
-      <Component {...pageProps} />
-    </UserProvider>
-  );
-}
-export default LessonsApp;
